refactor(mcp/client): Organize, simplify, and document the client module
This refactors the `crates/mcp/src/client` module to improve its organization, simplify its components, and add comprehensive documentation.

The key changes are:
- The health monitoring logic has been clarified. `HealthMonitor` is now a passive data store, and `McpClientManager` is solely responsible for the active monitoring loop.
- The HTTP transport has been reorganized. The complex Server-Sent Events (SSE) logic has been extracted from the main `http.rs` file into its own `http/sse.rs` submodule.
- The stdio transport has been improved. The `is_alive` and `close` methods now correctly manage the child process lifecycle.
- Comprehensive documentation has been added to all public modules, structs, traits, and functions in the `client` module.
