+++ crates/mcp/src/client/health.rs
@@ -1,133 +1,83 @@
 //! Health monitoring for MCP clients.
+//!
+//! This module provides the `HealthMonitor`, a thread-safe data store for tracking
+//! the health status of multiple MCP plugins. It is designed to be a passive
+//! component, holding health information that is updated by an external actor,
+//! such as the `McpClientManager`.
+//!
+//! ## Design
+//!
+//! - `HealthCheckResult`: A simple struct to represent the outcome of a single health check.
+//! - `HealthMonitor`: A container for the health status of all registered plugins. It
+//!   provides methods to register/unregister plugins, update their health, and query
+//!   their current status.
+//!
+//! The `HealthMonitor` itself does not perform any health checks; it only stores the
+//! results. The `McpClientManager` is responsible for scheduling and executing the
+//! health checks periodically.
 
 use crate::types::HealthStatus;
 use std::collections::HashMap;
 use std::sync::Arc;
-use std::time::{Duration, SystemTime};
 use tokio::sync::Mutex;
-use tokio::time::{interval, sleep};
-use tracing::{debug, warn};
+use tracing::debug;
 
-/// Health check result.
+/// Represents the result of a single health check.
 #[derive(Debug, Clone)]
 pub struct HealthCheckResult {
-    /// Whether the service is healthy.
+    /// Whether the service is considered healthy.
     pub healthy: bool,
 
-    /// Latency in milliseconds.
+    /// The latency of the health check in milliseconds, if successful.
     pub latency_ms: Option<u64>,
 
-    /// Error message if unhealthy.
+    /// An error message if the service is unhealthy.
     pub error: Option<String>,
 }
 
-/// Health monitor for tracking plugin health.
-#[derive(Clone)]
+/// A thread-safe data store for tracking the health of multiple plugins.
+///
+/// The `HealthMonitor` is a passive component that holds the health status for
+/// any number of plugins, identified by unique string names. It is intended to be
+/// updated by an active monitoring system, like the `McpClientManager`.
+#[derive(Clone, Debug)]
 pub struct HealthMonitor {
-    /// Health status for each plugin.
+    /// A map from plugin names to their current health status.
     health_status: Arc<Mutex<HashMap<String, HealthStatus>>>,
-
-    /// Health check interval.
-    check_interval: Duration,
-
-    /// Health check timeout.
-    check_timeout: Duration,
-
-    /// Whether monitoring is active.
-    monitoring: Arc<Mutex<bool>>,
 }
 
 impl HealthMonitor {
-    /// Create a new health monitor.
+    /// Creates a new, empty `HealthMonitor`.
     pub fn new() -> Self {
         Self {
             health_status: Arc::new(Mutex::new(HashMap::new())),
-            check_interval: Duration::from_secs(30),
-            check_timeout: Duration::from_secs(10),
-            monitoring: Arc::new(Mutex::new(false)),
         }
     }
 
-    /// Create a new health monitor with custom settings.
-    pub fn with_settings(check_interval: Duration, check_timeout: Duration) -> Self {
-        Self {
-            health_status: Arc::new(Mutex::new(HashMap::new())),
-            check_interval,
-            check_timeout,
-            monitoring: Arc::new(Mutex::new(false)),
-        }
-    }
-
-    /// Start health monitoring.
-    pub async fn start(&self) {
-        let mut monitoring = self.monitoring.lock().await;
-        if *monitoring {
-            return; // Already monitoring
-        }
-        *monitoring = true;
-        drop(monitoring);
-
-        let health_status = Arc::clone(&self.health_status);
-        let monitoring = Arc::clone(&self.monitoring);
-        let check_interval = self.check_interval;
-        let check_timeout = self.check_timeout;
-
-        tokio::spawn(async move {
-            let mut interval = interval(check_interval);
-
-            loop {
-                interval.tick().await;
-
-                // Check if we should still be monitoring
-                let should_monitor = {
-                    let monitoring = monitoring.lock().await;
-                    *monitoring
-                };
-
-                if !should_monitor {
-                    break;
-                }
-
-                // Perform health checks for all registered plugins
-                let plugin_names = {
-                    let status = health_status.lock().await;
-                    status.keys().cloned().collect::<Vec<_>>()
-                };
-
-                for plugin_name in plugin_names {
-                    if let Err(e) = Self::perform_health_check(&health_status, &plugin_name, check_timeout).await {
-                        warn!("Health check failed for plugin {}: {}", plugin_name, e);
-                    }
-                }
-            }
-        });
-
-        debug!("Health monitoring started");
-    }
-
-    /// Stop health monitoring.
-    pub async fn stop(&self) {
-        let mut monitoring = self.monitoring.lock().await;
-        *monitoring = false;
-        debug!("Health monitoring stopped");
-    }
-
-    /// Register a plugin for health monitoring.
+    /// Registers a new plugin for health monitoring.
+    ///
+    /// This will add the plugin to the store with a default, unhealthy status.
+    /// If the plugin is already registered, its status will be reset.
     pub async fn register_plugin(&self, plugin_name: String) {
-        let name_clone = plugin_name.clone();
         let mut status = self.health_status.lock().await;
-        status.insert(plugin_name, HealthStatus::new());
-        debug!("Registered plugin for health monitoring: {}", name_clone);
+        status.insert(plugin_name.clone(), HealthStatus::new());
+        debug!("Registered plugin for health monitoring: {}", plugin_name);
     }
 
-    /// Unregister a plugin from health monitoring.
+    /// Unregisters a plugin from health monitoring.
+    ///
+    /// If the plugin was registered, it is removed from the store.
     pub async fn unregister_plugin(&self, plugin_name: &str) {
         let mut status = self.health_status.lock().await;
-        status.remove(plugin_name);
-        debug!("Unregistered plugin from health monitoring: {}", plugin_name);
+        if status.remove(plugin_name).is_some() {
+            debug!("Unregistered plugin from health monitoring: {}", plugin_name);
+        }
     }
 
-    /// Update health status for a plugin.
+    /// Updates the health status for a registered plugin.
+    ///
+    /// If the plugin is found, its `HealthStatus` is updated based on the
+    /// provided `HealthCheckResult`.
     pub async fn update_health(&self, plugin_name: &str, result: HealthCheckResult) {
         let mut status = self.health_status.lock().await;
 
@@ -143,69 +93,18 @@ impl HealthMonitor {
         }
     }
 
-    /// Get health status for a plugin.
+    /// Retrieves the health status for a specific plugin.
+    ///
+    /// Returns `None` if the plugin is not registered.
     pub async fn get_health(&self, plugin_name: &str) -> Option<HealthStatus> {
         let status = self.health_status.lock().await;
         status.get(plugin_name).cloned()
     }
 
-    /// Get health status for all plugins.
+    /// Retrieves the health status for all registered plugins.
     pub async fn get_all_health(&self) -> HashMap<String, HealthStatus> {
         let status = self.health_status.lock().await;
         status.clone()
-    }
-
-    /// Perform a health check for a specific plugin.
-    async fn perform_health_check(
-        health_status: &Arc<Mutex<HashMap<String, HealthStatus>>>,
-        plugin_name: &str,
-        _timeout: Duration,
-    ) -> Result<(), String> {
-        // This is a placeholder implementation
-        // In practice, you would call the actual health check method
-        // for the specific plugin's transport
-
-        let start = SystemTime::now();
-
-        // Simulate a health check
-        sleep(Duration::from_millis(100)).await;
-
-        let latency = start.elapsed().unwrap_or_default().as_millis() as u64;
-
-        // Update the health status
-        let mut status = health_status.lock().await;
-        if let Some(health) = status.get_mut(plugin_name) {
-            health.mark_healthy();
-            health.handshake_latency = Some(latency);
-        }
-
-        Ok(())
-    }
-
-    /// Check if monitoring is active.
-    pub async fn is_monitoring(&self) -> bool {
-        let monitoring = self.monitoring.lock().await;
-        *monitoring
-    }
-
-    /// Get the check interval.
-    pub fn check_interval(&self) -> Duration {
-        self.check_interval
-    }
-
-    /// Get the check timeout.
-    pub fn check_timeout(&self) -> Duration {
-        self.check_timeout
-    }
-
-    /// Set the check interval.
-    pub fn set_check_interval(&mut self, interval: Duration) {
-        self.check_interval = interval;
-    }
-
-    /// Set the check timeout.
-    pub fn set_check_timeout(&mut self, timeout: Duration) {
-        self.check_timeout = timeout;
     }
 }
 
@@ -220,7 +119,7 @@ mod tests {
     use super::*;
 
     #[tokio::test]
-    async fn test_health_monitor() {
+    async fn test_health_monitor_registration() {
         let monitor = HealthMonitor::new();
 
         // Register a plugin
@@ -228,38 +127,60 @@ mod tests {
 
         // Check initial health
         let health = monitor.get_health("test-plugin").await;
-        assert!(health.is_some());
-        assert!(!health.unwrap().is_healthy());
+        assert!(health.is_some(), "Plugin should be registered");
+        assert!(!health.unwrap().is_healthy(), "Initial status should be unhealthy");
+    }
 
-        // Update health
-        let result = HealthCheckResult {
+    #[tokio::test]
+    async fn test_health_monitor_update() {
+        let monitor = HealthMonitor::new();
+        monitor.register_plugin("test-plugin".to_string()).await;
+
+        // Update health to healthy
+        let healthy_result = HealthCheckResult {
             healthy: true,
             latency_ms: Some(100),
             error: None,
         };
+        monitor.update_health("test-plugin", healthy_result).await;
+        let health = monitor.get_health("test-plugin").await.unwrap();
+        assert!(health.is_healthy(), "Status should be healthy");
+        assert_eq!(health.handshake_latency, Some(100));
 
-        monitor.update_health("test-plugin", result).await;
+        // Update health to unhealthy
+        let unhealthy_result = HealthCheckResult {
+            healthy: false,
+            latency_ms: None,
+            error: Some("Connection failed".to_string()),
+        };
+        monitor.update_health("test-plugin", unhealthy_result).await;
+        let health = monitor.get_health("test-plugin").await.unwrap();
+        assert!(!health.is_healthy(), "Status should be unhealthy");
+        assert_eq!(health.last_error, Some("Connection failed".to_string()));
+    }
 
-        let health = monitor.get_health("test-plugin").await;
-        assert!(health.unwrap().is_healthy());
+    #[tokio::test]
+    async fn test_health_monitor_unregistration() {
+        let monitor = HealthMonitor::new();
+        monitor.register_plugin("test-plugin".to_string()).await;
 
         // Unregister plugin
         monitor.unregister_plugin("test-plugin").await;
 
+        // Check health
         let health = monitor.get_health("test-plugin").await;
-        assert!(health.is_none());
+        assert!(health.is_none(), "Plugin should be unregistered");
     }
 
     #[tokio::test]
-    async fn test_health_monitor_start_stop() {
+    async fn test_get_all_health() {
         let monitor = HealthMonitor::new();
+        monitor.register_plugin("plugin-1".to_string()).await;
+        monitor.register_plugin("plugin-2".to_string()).await;
 
-        assert!(!monitor.is_monitoring().await);
-
-        monitor.start().await;
-        assert!(monitor.is_monitoring().await);
-
-        monitor.stop().await;
-        assert!(!monitor.is_monitoring().await);
+        let all_health = monitor.get_all_health().await;
+        assert_eq!(all_health.len(), 2);
+        assert!(all_health.contains_key("plugin-1"));
+        assert!(all_health.contains_key("plugin-2"));
     }
 }
